Move testing to jest removing security vuln
Several security vuln were found in the mocha dep chain and Jest has
been a breeze to work with so the easiest path to success was just to
convert.

The tests are now also themselves in TypeScript which did have an impact
on the test for updaterFor that was using some trickery that would not
work with type checking. Needless to say, this is now cleaned up and
using Jest mock functions.